Removed the connection lifespan tests since we have removed the feature. Renamed the test suite to 'monitor' since that what it now tests. Changed its assertion style from chai.assert to must.
