+++ package.json
@@ -37,6 +37,7 @@
     "scripts":
     {
         "test":       "./node_modules/.bin/plz display=dot test && ./node_modules/.bin/plz test-cov",
+        "test-simple": "./node_modules/.bin/mocha -R spec test/query.js test/units.js test/monitor.js test/discovery.js",
         "travis-cov": { "threshold": 85 }
     },
     "config":
